bot.py
import os, json, logging
from telegram import Update, InputFile
from telegram.ext import ApplicationBuilder, MessageHandler, filters, ContextTypes
from config import TOKEN, IMAGES_DIR, QUEST_FILE

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

GROUP_ID = -1001234567890

with open(QUEST_FILE, 'r', encoding='utf-8') as f:
    QUESTS = json.load(f)['quests']

async def start_quest(update: Update, context: ContextTypes.DEFAULT_TYPE):
    cmd = update.message.text.replace('/', '').lower()
    quest = QUESTS.get(cmd)
    if not quest:
        await update.message.reply_text('❌ Неизвестный квест.')
        return
    # здесь добавь свою логику для создания тем и отправки изображений

def create_app():
    app = ApplicationBuilder().token(TOKEN).build()
    # добавь свои обработчики
    return app