requirements.txt
python-telegram-bot==20.6
flask
python-dotenv==1.0.1